@@ -21,11 +21,8 @@ keywords = ["sequential learning", "reinforcement learning"]
 readme = "README.md"
 license = { file = "LICENSE" }
 
-requires-python = ">=3.7"
+requires-python = ">=3.10"
 classifiers = [
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3 :: Only",
@@ -59,7 +56,7 @@ write_to = "sabblium/_version.py"
 
 [tool.black]
 line-length = 88
-target-version = ['py37']
+target-version = ['py310']
 
 # ... other project metadata fields as specified in:
 #     https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
