@@ -1,12 +1,11 @@
-# coding=utf-8
+#  SaBBLium ― A Python library for building and simulating multi-agent systems.
 #
-# Copyright © Facebook, Inc. and its affiliates.
-# Copyright © Sorbonne University
+#  Copyright © Facebook, Inc. and its affiliates.
+#  Copyright © Sorbonne University.
 #
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
+#  This source code is licensed under the MIT license found in the
+#  LICENSE file in the root directory of this source tree.
 #
-from typing import Tuple
 
 import torch
 from gymnasium.utils import seeding
@@ -23,7 +22,8 @@ class ShuffledDatasetAgent(Agent):
         self,
         dataset: data.Dataset,
         batch_size: int,
-        output_names: Tuple[str, str] = ("x", "y"),
+        output_names: tuple[str, str] = ("x", "y"),
+        **kwargs,
     ):
         """Create the agent
         Args:
@@ -31,7 +31,7 @@ class ShuffledDatasetAgent(Agent):
             batch_size ([int]): The number of datapoints to write at each call
             output_names (tuple): The name of the variables. Default to ("x", "y").
         """
-        super().__init__()
+        super().__init__(**kwargs)
         self.output_names = output_names
         self.dataset: data.Dataset = dataset
         self.batch_size: int = batch_size
@@ -71,16 +71,18 @@ class DataLoaderAgent(Agent):
     If True, then no data have been written in the workspace since the reading of the dataset is terminated
     """
 
-    def __init__(self, dataloader, output_names: Tuple[str, str] = ("x", "y")):
+    def __init__(
+        self, dataloader, output_names: tuple[str, str] = ("x", "y"), **kwargs
+    ):
         """Create the agent based on a dataloader
         Args:
             dataloader ([DataLoader]): The underlying pytorch dataloader object
             output_names (tuple, optional): Names of the variable to write in the workspace. Defaults to ("x", "y").
         """
-        super().__init__()
+        super().__init__(**kwargs)
         self.dataloader = dataloader
         self.iter = iter(self.dataloader)
-        self.output_names: Tuple[str, str] = output_names
+        self.output_names: tuple[str, str] = output_names
         self._finished: bool = False
         self.ghost_params = torch.nn.Parameter(torch.randn(()))
 
@@ -92,6 +94,7 @@ class DataLoaderAgent(Agent):
         return self._finished
 
     def forward(self, **kwargs):
+        super().forward(**kwargs)
         try:
             output_values = next(self.iter)
         except StopIteration:
