@@ -1,17 +1,10 @@
-# coding=utf-8
+#  SaBBLium ― A Python library for building and simulating multi-agent systems.
 #
-# Copyright © Facebook, Inc. and its affiliates.
-# Copyright © Sorbonne University
+#  Copyright © Facebook, Inc. and its affiliates.
+#  Copyright © Sorbonne University.
 #
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
+#  This source code is licensed under the MIT license found in the
+#  LICENSE file in the root directory of this source tree.
 #
 
-from .dataloader import DataLoaderAgent, ShuffledDatasetAgent
-from .seeding import (
-    SeedableAgent,
-    SeedableAgentLast,
-    SeedableAgentMean,
-    SeedableAgentSum,
-)
 from .utils import Agents, CopyTemporalAgent, PrintAgent, TemporalAgent, EpisodesStopped
